SquareDanceReasoning

Documentation for SquareDanceReasoning.

Overview

Dancer represents a dancer. A dancer is identified by its couple_number and its gender.

The location and facing direction of a Dancer at a given time is represented by DancerState. See Coordinate System to learn how dancer location and facing direction are described.

See Motion Primitives for the simplest of operations for moving dancers around.

One of the goals of SquareDanceReasoning is to identify what formation the dancers are in. See Hierarchy of Supported Square Dance Formations for a list of the formations that SquareDanceReasoning can currently recognize.

Some of the reasoning, including formation recognition, is performed using a rule based expert system. See Hierarchy of Knowledge Base Rules for a list of the rules that are implemented.

To facilitate recognizing formations, the relative position predicates in_front_of, behind, left_of, and right_of are provided.

SquareDanceReasoning.COUPLE_DISTANCE

SquareDanceReasoning.DANCER_SYMBOLS_URI

SquareDanceReasoning.FULL_CIRCLE

SquareDanceReasoning.AfterEightChainOne

SquareDanceReasoning.AfterEightChainOneRule

SquareDanceReasoning.BackToBack

SquareDanceReasoning.BackToBackCouples

SquareDanceReasoning.BeforeEightChain

SquareDanceReasoning.BeforeEightChainRule

SquareDanceReasoning.Bounds

SquareDanceReasoning.BoxOfFour

SquareDanceReasoning.BoxOfFourRule

SquareDanceReasoning.CSSKeyframesAnimation

SquareDanceReasoning.CircleOfEight

SquareDanceReasoning.CircleOfEightFormationRule

SquareDanceReasoning.Collision

SquareDanceReasoning.CollisionRule

SquareDanceReasoning.ColumnOfFour

SquareDanceReasoning.ColumnOfFourRule

SquareDanceReasoning.CompletedDoublePassThru

SquareDanceReasoning.CompletedDoublePassThruRule

SquareDanceReasoning.Couple

SquareDanceReasoning.CoupleBoxRule

SquareDanceReasoning.Dancer

SquareDanceReasoning.DancerState

SquareDanceReasoning.EightDancerFormation

SquareDanceReasoning.FaceToFace

SquareDanceReasoning.FacingCouples

SquareDanceReasoning.FacingTandemCouples

SquareDanceReasoning.FacingTandemCouplesRule

SquareDanceReasoning.FourDancerFormation

SquareDanceReasoning.Gender

SquareDanceReasoning.Handedness

SquareDanceReasoning.LHBoxOfFour

SquareDanceReasoning.LHMiniWave

SquareDanceReasoning.LHWaveOfEight

SquareDanceReasoning.LHWaveOfFour

SquareDanceReasoning.LineOfFour

SquareDanceReasoning.LineOfFourRule

SquareDanceReasoning.MiniWave

SquareDanceReasoning.MoveApart

SquareDanceReasoning.MoveApart

SquareDanceReasoning.OriginalPartnerRule

SquareDanceReasoning.OriginalPartners

SquareDanceReasoning.PureSVGAnimation

SquareDanceReasoning.RHBoxOfFour

SquareDanceReasoning.RHMiniWave

SquareDanceReasoning.RHWaveOfEight

SquareDanceReasoning.RHWaveOfFour

SquareDanceReasoning.SDSquare

SquareDanceReasoning.SquareDanceFormation

SquareDanceReasoning.SquareDanceFormationRule

SquareDanceReasoning.SquareDanceRule

SquareDanceReasoning.SquareHasDancers

SquareDanceReasoning.SquaredSet

SquareDanceReasoning.SquaredSetFormationRule

SquareDanceReasoning.Tandem

SquareDanceReasoning.TandemCouples

SquareDanceReasoning.TemporalFact

SquareDanceReasoning.TimeBounds

SquareDanceReasoning.TwoDancerFormation

SquareDanceReasoning.TwoDancerFormationsRule

SquareDanceReasoning.TwoFacedLine

SquareDanceReasoning.TwoFacedLineRule

SquareDanceReasoning.WaveOfEight

SquareDanceReasoning.WaveOfEightRule

SquareDanceReasoning.WaveOfFour

SquareDanceReasoning.WaveOfFourRule

Base.isless

SquareDanceReasoning.animate

SquareDanceReasoning.backward

SquareDanceReasoning.behind

SquareDanceReasoning.breathe

SquareDanceReasoning.bump_out

SquareDanceReasoning.bump_out

SquareDanceReasoning.can_roll

SquareDanceReasoning.canonicalize

SquareDanceReasoning.center

SquareDanceReasoning.dancer_states

SquareDanceReasoning.dancer_symbol_uri

SquareDanceReasoning.direction

SquareDanceReasoning.direction

SquareDanceReasoning.direction_equal

SquareDanceReasoning.distance

SquareDanceReasoning.earliest

SquareDanceReasoning.elt

SquareDanceReasoning.encroached_on

SquareDanceReasoning.expand

SquareDanceReasoning.expand

SquareDanceReasoning.fixed

SquareDanceReasoning.formation_svg

SquareDanceReasoning.forward

SquareDanceReasoning.fraction

SquareDanceReasoning.grid_arrangement

SquareDanceReasoning.handedness

SquareDanceReasoning.history

SquareDanceReasoning.in_bounds

SquareDanceReasoning.in_front_of

SquareDanceReasoning.is_original_head

SquareDanceReasoning.is_original_side

SquareDanceReasoning.jitter

SquareDanceReasoning.latest_dancer_states

SquareDanceReasoning.left_of

SquareDanceReasoning.leftward

SquareDanceReasoning.make_kb

SquareDanceReasoning.make_kb

SquareDanceReasoning.make_square

SquareDanceReasoning.move_apart

SquareDanceReasoning.near

SquareDanceReasoning.opposite

SquareDanceReasoning.opposite

SquareDanceReasoning.other_dancers

SquareDanceReasoning.pass_by

SquareDanceReasoning.percentage

SquareDanceReasoning.quarter_left

SquareDanceReasoning.quarter_right

SquareDanceReasoning.revolve

SquareDanceReasoning.right_of

SquareDanceReasoning.rightward

SquareDanceReasoning.rotate

SquareDanceReasoning.square_up

SquareDanceReasoning.step_to_a_wave

SquareDanceReasoning.synchronize

SquareDanceReasoning.COUPLE_DISTANCE — Constant

COUPLE_DISTANCE is the distance between (the center reference points of) two dancers standing side by side, face to face, or back to back.

SquareDanceReasoning.DANCER_SYMBOLS_URI — Constant

DANCERSYMBOLSURI is a Dict that shows where to look for the dancer symbol definitions. This depends on what directory the formation HTML files are written to, which would determine whether the HTML will be accessed by a "local" URL, buy a GitHub Pages URL or whatever.

SquareDanceReasoning.FULL_CIRCLE — Constant

FULL_CIRCLE represents a change in direction of 360 degrees.

SquareDanceReasoning.AfterEightChainOne — Type

AfterEightChainOne is a formation of eight dancers (four couples) lined up, and is the starting formation for the call "trade by". It can result from doing a "centers pass thru" from a BeforeEightChain formation.

←→←→
←→←→

SquareDanceReasoning.AfterEightChainOneRule — Type

AfterEightChainOneRule is the rule for recognizing AfterEightChainOne formations.

SquareDanceReasoning.BackToBack — Type

BackToBack represents a formation of two dancers with their backs facing each other.

SquareDanceReasoning.BackToBackCouples — Type

BackToBackCouples is a formation that includes two Couple formations that have their backs to each other.

SquareDanceReasoning.BeforeEightChain — Type

BeforeEightChain is a formation of eight dancers (four couples) lined up, and is the starting formation for the "eight chain thru" call. It can result from doing "centers pass thru" from a FacingTandemCouples formation.

→←→←
→←→←

SquareDanceReasoning.BeforeEightChainRule — Type

BeforeEightChainRule is the rule for recognizing BeforeEightChain formations.

SquareDanceReasoning.Bounds — Type

Bounds(dss::Vector{DancerState};
                    margin = COUPLE_DISTANCE / 2)

represents the bounding rectangle surrounding the specified DancerStates. If margin is 0 then Bounds surrounds just the centers of the dancers.

By default, margin is COUPLE_DISTANCE / 2 so that Bounds describes the space actually occupied by the dancers.

SquareDanceReasoning.BoxOfFour — Type

BoxOfFour is the abstract supertype for RHBoxOfFour and LHBoxOfFour.

SquareDanceReasoning.BoxOfFourRule — Type

BoxOfFourRule is the rule for identifying right or left handed "box circulate" formations.

SquareDanceReasoning.CSSKeyframesAnimation — Type

CSSKeyfraneomesAnimation

As the method argument to animate, uses the CSS keyframes to animte SVG.

In my, and others's experience, can't get rotation to work.

SquareDanceReasoning.CircleOfEight — Type

CircleOfEight represents eight dancers in a circle facing inward.

SquareDanceReasoning.CircleOfEightFormationRule — Type

CircleOfEightFormationRule is the rule for recognizing eight dancers in a circle: CircleOfEight.

SquareDanceReasoning.Collision — Type

Collision notes that two dancers are occupying the same space.

SquareDanceReasoning.CollisionRule — Type

CollisionRule is a rule for detecting when two DancerStates occupy the same space. It asserts Collision.

SquareDanceReasoning.ColumnOfFour — Type

ColumnOfFour(lead::Tandem, tail::Tandem, centers::Tandem)

Represents a column of four dancers, each facing the back of the dancer in front of them.

SquareDanceReasoning.ColumnOfFourRule — Type

ColumnOfFourRule is the rule for identifying ColumnOfFour formations.

SquareDanceReasoning.CompletedDoublePassThru — Type

CompletedDoublePassThru is a formation of eight dancers (four couples) lined up, and is the ending formation for the call "double pass thru". It can result from doing a "centers pass thru" from a TradeBy formation.

←←→→
←←→→

SquareDanceReasoning.CompletedDoublePassThruRule — Type

CompletedDoublePassThruRule is the rule for rtecognizing CompletedDoublePassThru formations.

SquareDanceReasoning.Couple — Type

Couple(beau::DancerState, belle::DancerState)

Couple represents a formation of two dancers both facing the same direction.

SquareDanceReasoning.CoupleBoxRule — Type

CoupleBoxRule is the rule for identifying two couples arranged in a two by two box.

SquareDanceReasoning.Dancer — Type

Dancer(couple_number::Int, ::Gender)

Dancer represents a dancer.

SquareDanceReasoning.DancerState — Type

DancerState(dancer, time, direction, down, left)
DancerState(previoous::DancerSTate, time, direction, down, left)

represents the location and facing direction of a single dancer at a moment in time.

time is a number defining a temporal ordering. It could represent a number of beats, for example.

SquareDanceReasoning.EightDancerFormation — Type

EightDancerFormation is the abstract supertype of all square dance formations involving eight dancers.

SquareDanceReasoning.FaceToFace — Type

FaceToFace represents a formation of two dancers facing each other.

SquareDanceReasoning.FacingCouples — Type

FacingCouples is a formation that includes two Couple formations that are facing each other.

SquareDanceReasoning.FacingTandemCouples — Type

FacingTandemCouples is a formation of eight dancers (four couples) lined up, and is the starting formation for the "double pass thru" call.

→→←←
→→←←

SquareDanceReasoning.FacingTandemCouplesRule — Type

FacingTandemCouplesRule is the rule for recognizing the FacingTandemCouples formation.

SquareDanceReasoning.FourDancerFormation — Type

FourDancerFormation is the abstract supertype of all square dance formations involving four dancers.

SquareDanceReasoning.Gender — Type

Gender represents the gender of a dancer, which might be Guy, Gal or Unspecified.

Unspecified exists for when we want to emphasize gender agnosticism in a diagram.

Gender equality: Guy() == Guy(), Gal() == Gal(), otherwise not equal.

Use opposite to get the opposite gender.

SquareDanceReasoning.Handedness — Type

Handedness

square dance formations have a handedness, one of RightHanded(), LeftHanded() or NoHandedness().

opposite of a handedness returns the other handedness. NoHandednessis its own opposite.

SquareDanceReasoning.LHBoxOfFour — Type

LHBoxOfFour represents a left handed "box circulate" formation.

SquareDanceReasoning.LHMiniWave — Type

LHMiniWave represents a left handed wave of two dancers.

SquareDanceReasoning.LHWaveOfEight — Type

LHWaveOfEight represents a right handed wave of eight dancers.

SquareDanceReasoning.LHWaveOfFour — Type

LHWaveOfFour represents a left handed wave of four dancers.

SquareDanceReasoning.LineOfFour — Type

LineOfFour represents a line of four dancerrs all facing in the same direction.

SquareDanceReasoning.LineOfFourRule — Type

LineOfFourRule is the rule for iderntifying a LineOfFour formation.

SquareDanceReasoning.MiniWave — Type

MiniWave is the abstract supertype for all two dancer waves.

SquareDanceReasoning.MoveApart — Type

MoveApart(center, displacement)
MoveApart(collision::Collision, playmates)

struct MoveApart can be applied to a DancerState to determine how much it should be displaced to resolve collision.

center and displacement are two element Vectors.

playmates is a Vector of Vectors of DancerStates where the elements of each inner vector are working together.

SquareDanceReasoning.MoveApart — Method

(mp::MoveApart)(ds::DancerState)

For the given DancerState, determine how much mp should displace it.

SquareDanceReasoning.OriginalPartnerRule — Type

OriginalPartnerRule is a rule that identifies the original partners of a square dance set.

SquareDanceReasoning.OriginalPartners — Type

OriginalPartners(guy::Dancer, gal::Dancer)

OriginalPartners is a fact in the knowledge base that identifies the original partners in a square.

SquareDanceReasoning.PureSVGAnimation — Type

PureSVGAnimation

As the method argument to animate, use pure SVG animate animation. Various sources suggest that SVG animate is deprecated, but it works.

SquareDanceReasoning.RHBoxOfFour — Type

RHBoxOfFour represents a right handed "box circulate" formation.

SquareDanceReasoning.RHMiniWave — Type

RHMiniWave represents a right handed wave of two dancers.

SquareDanceReasoning.RHWaveOfEight — Type

RHWaveOfEight represents a right handed wave of eight dancers.

SquareDanceReasoning.RHWaveOfFour — Type

RHWaveOfFour represents a right handed wave of four dancers.

SquareDanceReasoning.SDSquare — Type

SDSquare(dancers)

SDSquare represents the dancers in a square.

make_square(number_of_dancers) will return an SDSquare.

SDSquare is a fact that can be asserted to the knowledge base to inform it that the dancers form a square.

other_dancers(::SDSquare, dancers) returns the set of Dancers that are in the SDSquare but are not in dancers.

SquareDanceReasoning.SquareDanceFormation — Type

SquareDanceFormation is the abstract supertype of all square dance formations.

SquareDanceReasoning.SquareDanceFormationRule — Type

SquareDanceFormationRule

the group for all rules relating to square dance formations.

SquareDanceReasoning.SquareDanceRule — Type

SquareDanceRule is the abstract supertype for all rules defined in this package.

SquareDanceReasoning.SquareHasDancers — Type

SquareHasDancers is a convenience rule for asserting the Dancers from a SDSquare.

SquareDanceReasoning.SquaredSet — Type

SquaredSet(couples1::FacingCouples, couples2::FacingCouples)

SquaredSet represents a squared set of eight dancers.

SquareDanceReasoning.SquaredSetFormationRule — Type

SquaredSetFormationRule is the rule for identifying a SquaredSet.

SquareDanceReasoning.Tandem — Type

Tandem(leaderLLDancerState, trailer::DancerState)

Tandem repreents a formation of two dancers where the trailer is facing the back of the leader.

SquareDanceReasoning.TandemCouples — Type

TandemCouples is a formation for two Couples in Tandem.

SquareDanceReasoning.TemporalFact — Type

TemporalFact

Abstract supertype for any facts that have a temporal dependence.

SquareDanceReasoning.TimeBounds — Type

TimeBounds()

Returns an empty TimeBounds interval.

SquareDanceReasoning.TwoDancerFormation — Type

TwoDancerFormation is the abstract supertype of all square dance formations involving two dancers.

SquareDanceReasoning.TwoDancerFormationsRule — Type

TwoDancerFormationsRule is the rule for identifying all two dancer formations: Couple, FaceToFace, BackToBack, Tandem, RHMiniWave, and LHMiniWave.

SquareDanceReasoning.TwoFacedLine — Type

TwoFacedLine represents a two faced line formation.

SquareDanceReasoning.TwoFacedLineRule — Type

TwoFacedLineRule is the rule for identifying TwoFacedLine formations.

SquareDanceReasoning.WaveOfEight — Type

WaveOfEight is the abstract supertype for right and left handed waves of eight dancers.

SquareDanceReasoning.WaveOfEightRule — Type

WaveOfEightRule is the rule for identifying waves of eight dancers: RHWaveOfEight and LHWaveOfEight.

SquareDanceReasoning.WaveOfFour — Type

WaveOfFour is the abstract supertype for right and left handed waves of four dancers.

SquareDanceReasoning.WaveOfFourRule — Type

WaveOfFourRule is the rule for identifying waves of four dancers: RHWaveOfFour and LHWaveOfFour.

Base.isless — Method

Base.isless(::Dancer, ::Dancer)::Bool

Provides a total ordering for dancers.

SquareDanceReasoning.animate — Method

animate(method::AnimationMethod, output_file, dancer_states, bpm)

Write an SVG animation of the dancer_states to output_file.

dancer_states should have the most recent DancerState for each Dancer. the previous property of each DancerState is used to determine the Dancer's motion.

bpm or "beats per minute" provides a time scale for the animation. Assuming each unit of time in a DancerState is a single beat, bpm is used to calculate the total duration of one cycle through the animation.

SquareDanceReasoning.backward — Method

backward(ds::DancerState, distance, time_delta)::DancerState

Move the Dancer identified by ds backward (based on ds.direction) the specified distance by returning a new dancer state.

SquareDanceReasoning.behind — Method

behind(focus::DancerState, other::DancerState)::Bool

returns true if other is behind focus.

SquareDanceReasoning.breathe — Method

breathe(collisions::Vector{Collision}, playmates::Vector{Vector{DancerState}}, everyone::Vector)::Vector{DancerStates}

Moves the dancers apart such that those in collisions no longer overlap. Those grouped together in playmates are moved together, and remain fixed relative to one another. everyone includes all DancerStates that are to be affected.

SquareDanceReasoning.bump_out — Method

bump_out(bounds::Bounds, amount)

Returns a new Bounds object that is expanded at each edge by amount.

SquareDanceReasoning.bump_out — Method

bump_out(bounds::Bounds)

Returns a new Bounds object that is expanded by COUPLE_DISTANCE / 2 on each edge so that instead of encompassing the centers of each Dancer it encompasses whole dancers.

SquareDanceReasoning.can_roll — Method

can_roll(ds::DancerState)

Determine whether the modifier "and roll" can be applied to the DancerState. If the dancer can roll then can_roll returns a non-zero rotation value.

This is not useful for "and roll as if you could".

SquareDanceReasoning.canonicalize — Method

canonicalize(direction)

canonicalizes the direction to be between 0.0 and 1.0.

SquareDanceReasoning.center — Method

center(dss)

returns the center of the specified DancerStates as a two element Vector of down and left coordinates.

SquareDanceReasoning.dancer_states — Function

dancer_states(formation)::Vector{DancerState}

Returns a list of the DancerStates in the formation, in no particular order.

SquareDanceReasoning.dancer_symbol_uri — Method

dancer_symbol_uri(html_destination)

Returns a relative URI from the specified HTML page to a location of the dancer SVG symbols file.

html_destination is the path where the formation HTML file is to be written.

SquareDanceReasoning.direction — Function

direction(formation)

If all of the dancers of formation are facing in the same direction then return that direction. Otherwise get a no such method error.

SquareDanceReasoning.direction — Method

direction(focus, other)

returns the direction that other is from the point of view of focus.

SquareDanceReasoning.direction_equal — Method

direction_equal(direction1, direction2)::Bool =

returns true if direction1 and direction2 are roughly equal.

SquareDanceReasoning.distance — Method

distance(p1, p2)

returns the distance between the two points represented by vectors.

SquareDanceReasoning.earliest — Method

earliest(ds::DancerState)

Returns the earliest DancerState in the specified DancerState's previous chain.

SquareDanceReasoning.elt — Method

elt(f, tagname::AbstractString, things...)
elt(tagname::AbstractString, things...)

Return an XML element. f is called with a single argument: either an XML.AbstractXMLNode or a Pair describing an XML attribute to be added to the resulting element.

SquareDanceReasoning.encroached_on — Method

encroached_on(formations, kb)

Returns true if any of the other contemporary DancerStates in the knowledge base are within the Bounds of the specified formations.

SquareDanceReasoning.expand — Method

expand(bounds::Bounds, ::DancerState)::Bounds

bounds is modified to encompass the additional DancerState.

SquareDanceReasoning.expand — Method

expand(tb::TimeBounds, ds::DancerState)::TimeBounds

expands tb to encompass the time of the specified DancerStates.

SquareDanceReasoning.fixed — Method

fixed(x)

truncates ludicrously small floating point numbers.

SquareDanceReasoning.formation_svg — Method

formation_svg(f::SquareDanceFormation, symbol_uri_base; id=nothing)

Returns an SVG element that illustrates the formation.

symbol_uri_base is a relative URL to the SVG symbols file. See dancer_symbol_uri.

If id is specified, it will be the XML Id of the drawing.

If margin is specified, its the additional space that surrounds the formation.

SquareDanceReasoning.forward — Method

forward(ds::DancerState, distance, time_delta)::DancerState

Move the Dancer identified by ds forward (based on ds.direction) the specified distance by returning a new dancer state.

SquareDanceReasoning.fraction — Method

fraction(tb::TimeBounds, t)

return where t falls within tb as a fraction. For t == tb.min the result would be 0. For t == tb.max the result would be 100.

SquareDanceReasoning.grid_arrangement — Method

 grid_arrangement(dancers::Vector{Dancer}, dancer_indices::Array{<:Integer, 2}, dancer_directions::Array{Rational, 2})

Returns a two dimensional array of DancerStates such that the dancers are arranges in a rectangular grid of the specified dimensions.

dancers is a vector of Dancers indexed by dancer_indices. A dancer index can be 0 to indicate that there is no dancer in that positriion.

dancer_indices and dancer_directions are Arrays of the same dimensions. This determiines the dimensions of the resulting grid.

Alternatively, dancer_directions can be a Vector of strings of the Unicode arrow characters ←, ↑, →, or ↓, with the space character serving as a placeholder where there is no dancer.

The two indices of each array correspond to the down and left coordinates respectively.

SquareDanceReasoning.handedness — Function

handedness(formation)

returns the handedness of a square dance formation: one of RightHanded(), LeftHanded() or NoHandedness()

SquareDanceReasoning.history — Method

history(f, ds::DancerState)

Calls f on ds and all of its previous DancerStates in chronological order.

SquareDanceReasoning.in_bounds — Method

in_bounds(bounds::Bounds, ds::DancerState)::Bool

Returns true if the specified DancerState it located within bounds.

in_bounds(Bounds([ds]), ds) == true

SquareDanceReasoning.in_front_of — Method

in_front_of(focus::DancerState, other::DancerState)::Bool

returns true if other is in front of focus.

SquareDanceReasoning.is_original_head — Method

is_original_head(::Dancer)::Bool

returns true if the dancer was originally in a head position.

SquareDanceReasoning.is_original_side — Method

is_original_side(::Dancer)::Bool

returns true if the dancer was originally in a side position.

SquareDanceReasoning.jitter — Method

jitter(ds::DancerState, time_delta)::DancerState

Adds noise to the position and facing direction of the DancerState. This allows us to test whether our formation recognition rules are sensitive to such noise.

SquareDanceReasoning.latest_dancer_states — Method

latest_dancer_states(kb)::Dict{Dancer, DancerState}

Return a Dict that provides the latest DancerState for each Dancer.

SquareDanceReasoning.left_of — Method

left_of(focus::DancerState, other::DancerState)::Bool

returns true if other is to the left of focus.

SquareDanceReasoning.leftward — Method

leftward(ds::DancerState, distance, time_delta)::DancerState

Move the Dancer identified by ds to the left (based on ds.direction) by the specified distance by returning a new dancer state.

SquareDanceReasoning.make_kb — Method

make_kb(kb::ReteRootNode)

Makes a copy of the knowledge base, but without any of the temporal facts.

SquareDanceReasoning.make_kb — Method

make_kb()

Creates SquareDanceReasoning knowledge base with no facts, but with all rules installed.

SquareDanceReasoning.make_square — Method

make_square(number_of_couples::Int)::SDSquare

Returns an SDSquare with the specified number of couples.

SquareDanceReasoning.move_apart — Method

move_apart(mps::Vector{MoveApart}, ds::DancerState)::DancerState

Applies all of the MoveApart displacements to the specified DancerState.

SquareDanceReasoning.near — Method

near(::DancerState, ::DancerState)

returns true if the two DancerStates are for different dancers and are close enough together (within DANCERNEARDISTANCE).

SquareDanceReasoning.opposite — Method

opposite(direction)

returns the direction opposite to the given one.

SquareDanceReasoning.opposite — Method

opposite(::Gender)::Gender

Returns the opposite Gender. Unspecified() is its own opposite.

SquareDanceReasoning.other_dancers — Method

other_dancers(square::SDSquare, dancers)

Returns the dancers from square that are not listed in the dancers argument.

SquareDanceReasoning.pass_by — Method

pass_by(mw::MiniWave, time_delta)::BackToBack

Dancers in a MiniWave pass by each other to end BackToBack.

SquareDanceReasoning.percentage — Method

percentage(tb::TimeBounds, t)

return where t falls within tb as a percentage. For t == tb.min the result would be 0. For t == tb.max the result would be 100.

SquareDanceReasoning.quarter_left — Method

quarter_left(direction)

returns the direction that's a quarter turn left from the given direction.

SquareDanceReasoning.quarter_right — Method

quarter_right(direction)

returns the direction that's a quarter turn right from the given direction.

SquareDanceReasoning.revolve — Method

revolve(ds::DancerState, center, new_direction, time_delta)::DancerState

Revolves the Dancer identified by ds around center (a two element Vector of down and left coordinates) until the Dancer's new facing direction is new_direction. A new DancerState is returned. time_delta is the duration of the revolution operation.

SquareDanceReasoning.right_of — Method

right_of(focus::DancerState, other::DancerState)::Bool

returns true if other is to the right of focus.

SquareDanceReasoning.rightward — Method

rightward(ds::DancerState, distance, time_delta)::DancerState

Move the Dancer identified by ds to the right (based on ds.direction) by the specified distance by returning a new dancer state.

SquareDanceReasoning.rotate — Method

rotate(ds::DancerState, rotation, time_delta)::DancerState

Rotates the dancer idntified by the DancerState in place by rotation.

SquareDanceReasoning.square_up — Method

square_up(dancers; initial_time = 0)

returns a list of DancerStates for the initial squared set.

SquareDanceReasoning.step_to_a_wave — Method

step_to_a_wave(f::FaceToFace, time_delta, h::Handedness)::MiniWave

The face to face dancers move up to make a right or left handed miniwave.

Breating should be done separately.

SquareDanceReasoning.synchronize — Method

synchronize(root::ReteRootNode)

Posts a Synchronized fact to the knowledge base corresponding to the latest time among all DancerStates in the knowledge base.

For and DancerState that is behind that time, a new DancerState is asserted for the latest time.

It is recommended that synchronize should be called after every call or each part of a call.