Definitions

SquareDanceReasoning.COUPLE_DISTANCE — Constant

COUPLE_DISTANCE is the distance between (the center reference points of) two dancers standing side by side, face to face, or back to back.

SquareDanceReasoning.FULL_CIRCLE — Constant

FULL_CIRCLE represents a change in direction of 360 degrees.

SquareDanceReasoning.AfterEightChainOne — Type

AfterEightChainOne is a formation of eight dancers (four couples) lined up, and is the starting formation for the call "trade by". It can result from doing a "centers pass thru" from a BeforeEightChain formation.

←→←→
←→←→

SquareDanceReasoning.AfterEightChainOneRule — Type

AfterEightChainOneRule is the rule for recognizing AfterEightChainOne formations.

SquareDanceReasoning.AllPresent — Type

AllPresent(::SDSquare)

AllPresent is a fact that is asserted to the knowledgebase by SDSquareHasAttendanceRule when all of the Dancers in an SDSquare are present.

It can be used as a trigger for rules that depend on all Dancers having an associated DancerState.

SquareDanceReasoning.AndRoll — Type

AndRoll(; role=EveryOne())

CallerLab Plus call.

Timing: CallerLab: 2.

SquareDanceReasoning.Attendance — Type

Attendance(::SDSquare)

Attendane is a fact in the knowledgebase that leeps track of how many DancerStates in a square are present.

This is necessary because tests like encroached_on can only provide a correct result if all of the DanecerStates are available for consideration.

SquareDanceReasoning.AttendanceRule — Type

AttendanceRule

AttendanceRule is the rule that updated a Attendance as new DancerStates are asserted to the nowledgebase, and ultimately asserts AllPresent.

SquareDanceReasoning.BackToBack — Type

BackToBack represents a formation of two dancers with their backs facing each other.

SquareDanceReasoning.BackToBackCouples — Type

BackToBackCouples is a formation that includes two Couple formations that have their backs to each other.

SquareDanceReasoning.Balance — Type

Balance(; role=Everyone(), time=2)

Implements the Balance square dance call, except that dancers just jiggle in place. Maybe it should be renamed _Identify.

SquareDanceReasoning.BeforeEightChain — Type

BeforeEightChain is a formation of eight dancers (four couples) lined up, and is the starting formation for the "eight chain thru" call. It can result from doing "centers pass thru" from a FacingTandemCouples formation.

→←→←
→←→←

SquareDanceReasoning.BeforeEightChainRule — Type

BeforeEightChainRule is the rule for recognizing BeforeEightChain formations.

SquareDanceReasoning.Bounds — Type

Bounds(dss::Vector{DancerState};
                    margin = COUPLE_DISTANCE / 2)

represents the bounding rectangle surrounding the specified DancerStates. If margin is 0 then Bounds surrounds just the centers of the dancers.

By default, margin is COUPLE_DISTANCE / 2 so that Bounds describes the space actually occupied by the dancers.

SquareDanceReasoning.BoxOfFour — Type

BoxOfFour is the abstract supertype for RHBoxOfFour and LHBoxOfFour.

SquareDanceReasoning.BoxOfFourRule — Type

BoxOfFourRule is the rule for identifying right or left handed "box circulate" formations.

SquareDanceReasoning.CSSKeyframesAnimation — Type

CSSKeyfraneomesAnimation

As the method argument to animate, uses the CSS keyframes to animte SVG.

In my, and others's experience, can't get rotation to work.

SquareDanceReasoning.CallEngineDebugContext — Type

CallEngineDebugContext(source_location, relpath, token)

A CallEngineDebugContext can be bassed to some functions via the dbgctx keyword argument to provide debugging support.

Do not use the constructor directly, instead use @CallEngineDebugContext, which wil automatically provide a source location.

SquareDanceReasoning.CallSchedule — Type

CallSchedule()
CallSchedule(now)

Creates and returns an empty square dance call schedule, either starting at time 0 or at the specified time.

Use schedule to add a new entry to the schedule.

SquareDanceReasoning.CanDoCall — Type

CanDoCall(preference, call::SquareDanceCall, formation::SquareDanceFormation)

CanDoCall represents that call can be performed from formation, and doing so has the specified preference.

SquareDanceReasoning.CircleOfEight — Type

CircleOfEight represents eight dancers in a circle facing inward.

SquareDanceReasoning.CircleOfEightFormationRule — Type

CircleOfEightFormationRule is the rule for recognizing eight dancers in a circle: CircleOfEight.

SquareDanceReasoning.Collision — Type

Collision notes that two dancers are occupying the same space.

SquareDanceReasoning.ColumnOfFour — Type

ColumnOfFour(lead::Tandem, tail::Tandem, centers::Tandem)

Represents a column of four dancers, each facing the back of the dancer in front of them.

SquareDanceReasoning.ColumnOfFourRule — Type

ColumnOfFourRule is the rule for identifying ColumnOfFour formations.

SquareDanceReasoning.CompletedDoublePassThru — Type

CompletedDoublePassThru is a formation of eight dancers (four couples) lined up, and is the ending formation for the call "double pass thru". It can result from doing a "centers pass thru" from a TradeBy formation.

←←→→
←←→→

SquareDanceReasoning.CompletedDoublePassThruRule — Type

CompletedDoublePassThruRule is the rule for rtecognizing CompletedDoublePassThru formations.

SquareDanceReasoning.Couple — Type

Couple(beau::DancerState, belle::DancerState)

Couple represents a formation of two dancers both facing the same direction.

SquareDanceReasoning.CoupleBoxRule — Type

CoupleBoxRule is the rule for identifying two couples arranged in a two by two box.

SquareDanceReasoning.Dancer — Type

Dancer(couple_number::Int, ::Gender)

Dancer represents a dancer.

SquareDanceReasoning.DancerState — Type

DancerState(dancer, time, direction, down, left)
DancerState(previoous::DancerSTate, time, direction, down, left)

represents the location and facing direction of a single dancer at a moment in time.

time is a number defining a temporal ordering. It could represent a number of beats, for example.

DancerState is also the single dancer SquareDanceFormation.

SquareDanceReasoning.Dosado — Type

Dosado(; role=Everyone(), handedness=RightHanded())

CallerLab Basic1 call.

Timing: Callerlab: 6, unless coming from and returnuing to a squared set, in which case 8.

SquareDanceReasoning.EightDancerFormation — Type

EightDancerFormation is the abstract supertype of all square dance formations involving eight dancers.

SquareDanceReasoning.FaceLeft — Type

FaceLeft(; role=Everyone(), time=2)

CallerLab Basic 1 square dance call causing dancers to turn 1/4 to their left. The timing defaults to 2 since, according to Taminations, two beats is the duration for QuarterIn and QuarterOut.

Timing: CallerLab does not soecify timing. Assume 1. Can be specified as a parameter.

SquareDanceReasoning.FaceRight — Type

FaceRight(; role=Everyone(), time=2)

CallerLab Basic 1 square dance call causing dancers to turn 1/4 to their right. The timing defaults to 2 since, according to Taminations, two beats is the duration for QuarterIn and QuarterOut.

Timing: CallerLab does not soecify timing. Assume 1. Can be specified as a parameter.

SquareDanceReasoning.FaceToFace — Type

FaceToFace represents a formation of two dancers facing each other.

SquareDanceReasoning.FacingCouples — Type

FacingCouples is a formation that includes two Couple formations that are facing each other.

SquareDanceReasoning.FacingTandemCouples — Type

FacingTandemCouples is a formation of eight dancers (four couples) lined up, and is the starting formation for the "double pass thru" call.

→→←←
→→←←

SquareDanceReasoning.FacingTandemCouplesRule — Type

FacingTandemCouplesRule is the rule for recognizing the FacingTandemCouples formation.

SquareDanceReasoning.FourDancerFormation — Type

FourDancerFormation is the abstract supertype of all square dance formations involving four dancers.

SquareDanceReasoning.Gal — Type

Gal()

The gal gender, represented by a circle in our SVG diagrams.

SquareDanceReasoning.Gender — Type

Gender represents the gender of a dancer, which might be Guy, Gal or Unspecified.

Unspecified exists for when we want to emphasize gender agnosticism in a diagram.

Gender equality: Guy() == Guy(), Gal() == Gal(), otherwise not equal.

Use opposite to get the opposite gender.

SquareDanceReasoning.Guy — Type

Guy()

The guy gender, represented by a square in our SVG diagrams.

SquareDanceReasoning.Handedness — Type

Handedness

square dance formations have a handedness, one of RightHanded(), LeftHanded() or NoHandedness().

opposite of a handedness returns the other handedness. NoHandednessis its own opposite.

SquareDanceReasoning.Hinge — Type

Hinge(; role=Everyone(), tile=2)

CallerLab Mainstream call.

Timing: CallerLab: 2.

SquareDanceReasoning.LHBoxOfFour — Type

LHBoxOfFour represents a left handed "box circulate" formation.

SquareDanceReasoning.LHMiniWave — Type

LHMiniWave represents a left handed wave of two dancers.

SquareDanceReasoning.LHWaveOfEight — Type

LHWaveOfEight represents a right handed wave of eight dancers.

SquareDanceReasoning.LHWaveOfFour — Type

LHWaveOfFour represents a left handed wave of four dancers.

SquareDanceReasoning.LineOfFour — Type

LineOfFour represents a line of four dancerrs all facing in the same direction.

SquareDanceReasoning.LineOfFourRule — Type

LineOfFourRule is the rule for iderntifying a LineOfFour formation.

SquareDanceReasoning.MiniWave — Type

MiniWave is the abstract supertype for all two dancer waves.

SquareDanceReasoning.ObverseRole — Type

ObverseRole(::role)

For roles like CoupleNumbers or DesignatedDancers we can't determine the obverse until we know all of the dancers involved. TRhis servesd as a "place holder" until we can determine which dancers have the role that is obverse to the specified Role.

SquareDanceReasoning.OriginalPartnerRule — Type

OriginalPartnerRule is a rule that identifies the original partners of a square dance set.

SquareDanceReasoning.OriginalPartners — Type

OriginalPartners(guy::Dancer, gal::Dancer)

OriginalPartners is a fact in the knowledge base that identifies the original partners in a square.

SquareDanceReasoning.PassThru — Type

PassThru(; role=Everyone())

CallerLab Basic 1 call.

Timing: CallerLab: 2.

PassThru is only proper from FaceToFace or from RHMiniWave (because of the "Icean Wave Rule"). For LHMiniWave, use StepThru.

SquareDanceReasoning.PullBy — Type

PullBy(; role=Everyone(), handedness=RightHanded())

CallerLab Basic 1 call.

Timing: CallerLab doesn't specify timing, but since the timing for PassThru is 2, assume the same for PullBy.

SquareDanceReasoning.PureSVGAnimation — Type

PureSVGAnimation

As the method argument to animate, use pure SVG animate animation. Various sources suggest that SVG animate is deprecated, but it works.

SquareDanceReasoning.QuarterIn — Type

QuarterIn(; role=Everyone())

CalllerLab Advanced 1 call.

Timing: CallerLab: 2.

SquareDanceReasoning.QuarterOut — Type

QuarterOut(; role=Everyone())

CalllerLab Advanced 1 call.

Timing: CallerLab: 2.

SquareDanceReasoning.RHBoxOfFour — Type

RHBoxOfFour represents a right handed "box circulate" formation.

SquareDanceReasoning.RHMiniWave — Type

RHMiniWave represents a right handed wave of two dancers.

SquareDanceReasoning.RHWaveOfEight — Type

RHWaveOfEight represents a right handed wave of eight dancers.

SquareDanceReasoning.RHWaveOfFour — Type

RHWaveOfFour represents a right handed wave of four dancers.

SquareDanceReasoning.Respirator — Type

Respirator(::Vector{DancerState))

Respirator is a device for assisting breating. It performbs the bookkeeping associed with tracking the current location of each dancer as adjusted by the breathing algorithm without creating new DancerStates.

SquareDanceReasoning.SDSquare — Type

SDSquare(dancers)

SDSquare represents the dancers in a square.

make_square(number_of_dancers) will return an SDSquare.

SDSquare is a fact that can be asserted to the knowledge base to inform it that the dancers form a square.

other_dancers(::SDSquare, dancers) returns the set of Dancers that are in the SDSquare but are not in dancers.

SquareDanceReasoning.SDSquareHasAttendanceRule — Type

SquareDanceFormationRule

A rule that ensures that there is an Attendance fact for every SDSquare fact.

SquareDanceReasoning.SquareDanceCall — Type

SquareDanceCall

For each call, a struct is defined that is a subtype of SquareDanceCall.

For each call and each formation it can be done from, a can_do_from method should be defined. If the "Ocean Wave Rule" applies, then a can_do_from method should be defined for that case.

For each call and each formation it can be performed from, either a perform or an expand_parts method should be defined.

SquareDanceReasoning.SquareDanceFormation — Type

SquareDanceFormation is the abstract supertype of all square dance formations.

SquareDanceReasoning.SquareDanceFormationRule — Type

SquareDanceFormationRule

the group for all rules relating to square dance formations.

SquareDanceReasoning.SquareDanceRule — Type

SquareDanceRule is the abstract supertype for all rules defined in this package.

SquareDanceReasoning.SquareHasDancers — Type

SquareHasDancers is a convenience rule for asserting the Dancers from a SDSquare.

SquareDanceReasoning.SquareThru — Type

SquareThru(; role, handedness, count)

CallerLab Basic 1 square dance call.

Timing: count=1: 2, count=2: 5, count=3: 7 or 8, count=4: 10.

SquareDanceReasoning.SquaredSet — Type

SquaredSet(couples1::FacingCouples, couples2::FacingCouples)

SquaredSet represents a squared set of eight dancers.

SquareDanceReasoning.SquaredSetFormationRule — Type

SquaredSetFormationRule is the rule for identifying a SquaredSet.

SquareDanceReasoning.Star — Type

Star represents four dancers holding either right or left hands together in their center. All are facing either clockwise or counterclockwise.

SquareDanceReasoning.StarRule — Type

StarRule is the rule for recognizing right or left handed stars.

SquareDanceReasoning.StepThru — Type

StepThru(; role=Everyone())

CallerLab Basic 1 square dance call that goes from MiniWave to BackToBack.

Timing: CallerLab doesn't spoecify a timing, but since the specified timing for PassThru is 2 and StepThro must be smpler, assume 1.

SquareDanceReasoning.StepToAWave — Type

StepToAWave(; role=Everyone(), handedness=RightHanded())

CallerLab Basic 2 square dance call that goes from FaceToFace to a MiniWave of the specified handedness. The first half of PassThru.

Timing: 2.

SquareDanceReasoning.Tandem — Type

Tandem(leaderLLDancerState, trailer::DancerState)

Tandem repreents a formation of two dancers where the trailer is facing the back of the leader.

SquareDanceReasoning.TandemCouples — Type

TandemCouples is a formation for two Couples in Tandem.

SquareDanceReasoning.TemporalFact — Type

TemporalFact

Abstract supertype for any facts that have a temporal dependence.

SquareDanceReasoning.TimeBounds — Type

TimeBounds()

Returns an empty TimeBounds interval.

SquareDanceReasoning.TwoDancerFormation — Type

TwoDancerFormation is the abstract supertype of all square dance formations involving two dancers.

SquareDanceReasoning.TwoDancerFormationsRule — Type

TwoDancerFormationsRule

TwoDancerFormationsRule is the rule for identifying all two dancer formations: Couple, FaceToFace, BackToBack, Tandem, RHMiniWave, and LHMiniWave.

SquareDanceReasoning.TwoFacedLine — Type

TwoFacedLine represents a two faced line formation.

SquareDanceReasoning.TwoFacedLineRule — Type

TwoFacedLineRule is the rule for identifying TwoFacedLine formations.

SquareDanceReasoning.UTurnBack — Type

UTurnBack(; role=Everyone())

CallerLab Basic1 call.

Timing: CallerLab: 2.

SquareDanceReasoning.Unspecified — Type

Unspecified

Unspecified gender is provided just for gender neutral diagraming, where it appears as a square with rounded corners. It is not supported widely through this codebase. For example original partner and original corner are do not support it.

Note that

Unspecified() != Unspecified()

and

opposite(Unspecified())

returns Unspecified().

SquareDanceReasoning.WaveOfEight — Type

WaveOfEight is the abstract supertype for right and left handed waves of eight dancers.

SquareDanceReasoning.WaveOfEightRule — Type

WaveOfEightRule is the rule for identifying waves of eight dancers: RHWaveOfEight and LHWaveOfEight.

SquareDanceReasoning.WaveOfFour — Type

WaveOfFour is the abstract supertype for right and left handed waves of four dancers.

SquareDanceReasoning.WaveOfFourRule — Type

WaveOfFourRule is the rule for identifying waves of four dancers: RHWaveOfFour and LHWaveOfFour.

SquareDanceReasoning._BackToAWave — Type

_BackToAWave, handedness(; role=Everyone(), handedness=RightHanded())

Primitive square dance call that goes from BackToBack to a MiniWave of the specified handedness. The third quarter of Dosado.

Timing: 1.

SquareDanceReasoning._GenderedRoll — Type

_GenderedRoll(; role=Everyone(), time=2)

The second part of calls like StarThru and SlideThru. Guy turn one quarter to the right, Gal turn one quarter to the left.

Timing: as specified in the parameter. Defaults to 2. CallerLab says the timing for StarThru is 4, so thisseems reasonable.

SquareDanceReasoning._Rest — Type

_Rest(; role=Everyone(), time)

Primitive square dance call causing dancers to rest in place for the specified time.

Timing: as specified in the parameter.

SquareDanceReasoning._UnStepToAWave — Type

_UnStepToAWave(; role=Everyone())

Primitive square dance call that goes from a MiniWave to FaceToFace. The first half of PassThru.

Timing: 2. Persimably should have the same timing as StepToAWave

Base.isless — Method

Base.isless(::Dancer, ::Dancer)::Bool

Provides a total ordering for dancers.

SquareDanceReasoning.advance_schedule_by — Method

advance_schedule_by(sched::CallSchedule, delta)

Move every entry in the schedule forward by delta.

SquareDanceReasoning.animate — Method

animate(method::AnimationMethod, output_file, dancer_states, bpm)

Returns an SVG document fragment that animates the motion of the Dancers in dancer_states.

If output_file ends in ".svg" then the document fragment is written to that file. output_file is still needed so that the relative path to the dancer symbols file canbe determined.

dancer_states should have the most recent DancerState for each Dancer. the previous property of each DancerState is used to determine the Dancer's motion.

bpm or "beats per minute" provides a time scale for the animation. Assuming each unit of time in a DancerState is a single beat, bpm is used to calculate the total duration of one cycle through the animation.

SquareDanceReasoning.back_to_a_wave — Method

back_to_a_wave(f::BackToBack, time_delta, h::Union{RightHanded, LeftHanded})::FaceToFace

The back to back dancers back up to firm a miniwave of the specified handedness.

Breating should be done separately.

SquareDanceReasoning.backward — Method

backward(ds::DancerState, distance, time_delta)::DancerState

Move the Dancer identified by ds backward (based on ds.direction) the specified distance by returning a new dancer state.

SquareDanceReasoning.behind — Method

behind(focus::DancerState, other::DancerState)::Bool

returns true if other is behind focus.

SquareDanceReasoning.breathe — Method

breathe(playmates::Vector{TwoDancerFormation}, everyone::Vector{DancerState})::Vector{DancerState}

Moves the dancers apart such that those that have collided no longer overlap.

playmates is used to inform what direction colliders should be moved in such that playmates are not separated.

everyone includes all DancerStates that are to be affected.

SquareDanceReasoning.bump_out — Method

bump_out(bounds::Bounds, amount)

Returns a new Bounds object that is expanded at each edge by amount.

SquareDanceReasoning.bump_out — Method

bump_out(bounds::Bounds)

Returns a new Bounds object that is expanded by COUPLE_DISTANCE / 2 on each edge so that instead of encompassing the centers of each Dancer it encompasses whole dancers.

SquareDanceReasoning.can_do_from — Method

candofrom(::SquareDanceCall, ::SquareDanceFormation)::Int

Determins if the specified call can be performed from the specified formation. A return value of 0 means the call is not appropriate (or not currently supported). Otherwise the return value is a preference level with a higher value indicating more preferable. For example UTurnBack from Couple is more preferable to UTurnBack from DancerState, even though both would be applicable.

SquareDanceReasoning.can_roll — Method

can_roll(ds::DancerState)

Determine whether the modifier "and roll" can be applied to the DancerState. If the dancer can roll then can_roll returns a signed rotation value.

This is not useful for "and roll as if you could".

SquareDanceReasoning.canonicalize — Method

canonicalize(direction)

canonicalizes the direction to be between 0 and 1.

SquareDanceReasoning.canonicalize_signed — Method

canonicalize_signed(direction)

canonicalizes the direction to be between -1//2 and 1//2.

SquareDanceReasoning.center — Method

center(dss)

returns the center of the specified DancerStates as a two element Vector of down and left coordinates.

SquareDanceReasoning.collateral_file_relpath — Method

collateral_file_relpath(resource_name, html_file_destination)

Returns the path to the named resource in the environment of html_file_destination.

resource_name is one of the keys of HTML_COLLATERAL_FILES.

html_file_destination is the absolute path of the XML or HTML file that is referring to the resource.

SquareDanceReasoning.collisions — Method

collisions(::Respirator)

Returns an iterator over the current collisions in the Respirator.

SquareDanceReasoning.copy_html_collateral_files — Method

copy_html_collateral_files()

Copies each of the files identified in HTMLCOLLATERALFILES to each of the directories listed in TARGET_LOCATIONS.

SquareDanceReasoning writes XML and HTML files to various locations. Some of those locations are only accessible in certain circumstances. For example, The package source code hierarchy is not available to documentation foles delivered to the gh-pages branch. copy_html_collateral_files copies the files to where they might be needed. collateral_file_relpath is used to file the most appropriate

Use collateral_file_relpath to get the best reference location for a resource file.

SquareDanceReasoning.current_location — Method

current_location(r::Respirator, d::Dancer)

returns the current location of the Dancer as recorded in the Respirator.

SquareDanceReasoning.dancer_states — Function

dancer_states(formation)::Vector{DancerState}

Returns a list of the DancerStates in the formation, in no particular order.

SquareDanceReasoning.direction — Function

direction(formation)

If all of the dancers of formation are facing in the same direction then return that direction. Otherwise get a no such method error.

SquareDanceReasoning.direction — Method

direction(focus, other)

returns the absolute direction that other is from the point of view of focus. Either argument can be a DancerState or a localtion represented as a two element Vector.

SquareDanceReasoning.distance — Method

distance(p1, p2)

returns the distance between the two points represented by vectors.

SquareDanceReasoning.do_call — Method

do_call(kb::ReteRootNode, call::SquareDanceCall; dbgctx = nothing)::ReteRootNode

Perform the specified square dance call and retrn an updated knowledgebase.

SquareDanceReasoning.earliest — Method

earliest(ds::DancerState)

Returns the earliest DancerState in the specified DancerState's previous chain.

SquareDanceReasoning.elt — Method

elt(f, tagname::AbstractString, things...)
elt(tagname::AbstractString, things...)

Return an XML element. f is called with a single argument: either an XML.AbstractXMLNode or a Pair describing an XML attribute to be added to the resulting element.

SquareDanceReasoning.encroached_on — Method

encroached_on(formations, kb)

Returns true if any of the other contemporary DancerStates in the knowledge base are within the Bounds of the specified formations.

SquareDanceReasoning.expand — Method

expand(bounds::Bounds, ::DancerState)::Bounds

bounds is modified to encompass the additional DancerState.

SquareDanceReasoning.expand — Method

expand(tb::TimeBounds, ds::DancerState)::TimeBounds

expands tb to encompass the time of the specified DancerStates.

SquareDanceReasoning.expand_parts — Method

expand_parts(::SquareDanceCall, ::SquareDanceFormation)

Returns either the call itself, if it has no expansion, or a vector of Tuple{relative_time, part} containing one Tuple for each part. relative_time is the time (relative to the start of the call) at which the corresonding part should start.

SquareDanceReasoning.find_collisions — Method

find_collisions(dss::Vector{DancerState})::Vector{Collision}

Find the DancerStates that have collided with each other.

SquareDanceReasoning.fixed — Method

fixed(x)

truncates ludicrously small floating point numbers.

SquareDanceReasoning.formation_svg — Method

formation_svg(f::SquareDanceFormation, symbol_uri_base; id=nothing)

Returns an SVG element that illustrates the formation.

symbol_uri_base is a relative URL to the SVG symbols file. See collateral_file_relpath.

If id is specified, it will be the XML Id of the drawing.

If margin is specified, its the additional space that surrounds the formation.

SquareDanceReasoning.forward — Method

forward(ds::DancerState, distance, time_delta)::DancerState

Move the Dancer identified by ds forward (based on ds.direction) the specified distance by returning a new dancer state.

SquareDanceReasoning.fraction — Method

fraction(tb::TimeBounds, t)

return where t falls within tb as a fraction. For t == tb.min the result would be 0. For t == tb.max the result would be 100.

SquareDanceReasoning.grid_arrangement — Method

 grid_arrangement(dancers::Vector{Dancer}, dancer_indices::Array{<:Integer, 2}, dancer_directions::Array{Rational, 2})

Returns a two dimensional array of DancerStates such that the dancers are arranged in a rectangular grid of the specified dimensions.

dancers is a vector of Dancers indexed by dancer_indices. A dancer index can be 0 to indicate that there is no dancer in that positriion.

dancer_indices and dancer_directions are Arrays of the same dimensions. This determiines the dimensions of the resulting grid.

Alternatively, dancer_directions can be a Vector of strings of the Unicode arrow characters ←, ↑, →, or ↓, with the space character serving as a placeholder where there is no dancer.

The two indices of each array correspond to the down and left coordinates respectively.

SquareDanceReasoning.handedness — Function

handedness(formation)

returns the handedness of a square dance formation: one of RightHanded(), LeftHanded() or NoHandedness()

SquareDanceReasoning.history — Method

history(f, ds::DancerState)

Calls f on ds and all of its previous DancerStates in chronological order.

SquareDanceReasoning.in_bounds — Method

in_bounds(bounds::Bounds, ds::DancerState)::Bool

Returns true if the specified DancerState it located within bounds.

in_bounds(Bounds([ds]), ds) == true

SquareDanceReasoning.in_front_of — Method

in_front_of(focus::DancerState, other::DancerState)::Bool

returns true if other is in front of focus.

SquareDanceReasoning.is_original_head — Method

is_original_head(::Dancer)::Bool

returns true if the dancer was originally in a head position.

SquareDanceReasoning.is_original_side — Method

is_original_side(::Dancer)::Bool

returns true if the dancer was originally in a side position.

SquareDanceReasoning.jitter — Method

jitter(ds::DancerState, time_delta)::DancerState

Adds noise to the position and facing direction of the DancerState. This allows us to test whether our formation recognition rules are sensitive to such noise.

SquareDanceReasoning.latest_dancer_states — Method

latest_dancer_states(kb)::Dict{Dancer, DancerState}

Return a Dict that provides the latest DancerState for each Dancer.

SquareDanceReasoning.left_of — Method

left_of(focus::DancerState, other::DancerState)::Bool

returns true if other is to the left of focus.

SquareDanceReasoning.leftward — Method

leftward(ds::DancerState, distance, time_delta)::DancerState

Move the Dancer identified by ds to the left (based on ds.direction) by the specified distance by returning a new dancer state.

SquareDanceReasoning.make_kb — Method

make_kb(kb::ReteRootNode)

Makes a copy of the knowledge base, but without any TemporalFacts.

SquareDanceReasoning.make_kb — Method

make_kb()

Creates SquareDanceReasoning knowledge base with no facts, but with all rules installed.

SquareDanceReasoning.make_showcase_index_file — Method

make_showcase_index_file()

Writes an index file for the choreography Showcase docs directory.

SquareDanceReasoning.make_square — Method

make_square(number_of_couples::Int)::SDSquare

Returns an SDSquare with the specified number of couples.

SquareDanceReasoning.near — Method

near(::DancerState, ::DancerState)

returns true if the two DancerStates are for different dancers and are close enough together (within DANCERNEARDISTANCE).

SquareDanceReasoning.obverse — Function

obverse(::Role)::Role

Each role has an obverse role. (I didn't want to overload opposite). At a given time a dancer either has role or role's obverse.

SquareDanceReasoning.opposite — Method

opposite(direction)

returns the direction opposite to the given one.

SquareDanceReasoning.opposite — Method

opposite(::Gender)::Gender

Returns the opposite Gender. Unspecified() is its own opposite.

SquareDanceReasoning.other_dancers — Method

other_dancers(square::SDSquare, dancers)

Returns the dancers from square that are not listed in the dancers argument.

SquareDanceReasoning.pass_by — Method

pass_by(mw::MiniWave, time_delta)::BackToBack

Dancers in a MiniWave pass by each other to end BackToBack.

SquareDanceReasoning.percentage — Method

percentage(tb::TimeBounds, t)

return where t falls within tb as a percentage. For t == tb.min the result would be 0. For t == tb.max the result would be 100.

SquareDanceReasoning.perform — Function

perform(::SquareDanceCall, ::SquareDanceFormation, ::ReteRootNode)::SquareDanceFormation

Performs the call on the specified formation and returns the new formation.

SquareDanceReasoning.playmate — Function

playmate(ds::Dancer, f::TwoDancerFormation)::Union{Nothing, Dancer}

If the Dancer is part of the formation then return the other Dancer in the formation, otherwise return nothing.

SquareDanceReasoning.quarter_left — Method

quarter_left(direction)

returns the direction that's a quarter turn left from the given direction.

SquareDanceReasoning.quarter_right — Method

quarter_right(direction)

returns the direction that's a quarter turn right from the given direction.

SquareDanceReasoning.restricted_to — Method

restricted_to(call)::Role

Returns the role that call has been restricted to.

Each subtype of SquareDanceCall must have either a role field or a restricted_to method.

SquareDanceReasoning.revolve — Method

revolve(ds::DancerState, center, new_direction, time_delta)::DancerState

Revolves the Dancer identified by ds around center (a two element Vector of down and left coordinates) until the Dancer's new facing direction is new_direction. A new DancerState is returned. time_delta is the duration of the revolution operation.

SquareDanceReasoning.right_of — Method

right_of(focus::DancerState, other::DancerState)::Bool

returns true if other is to the right of focus.

SquareDanceReasoning.rightward — Method

rightward(ds::DancerState, distance, time_delta)::DancerState

Move the Dancer identified by ds to the right (based on ds.direction) by the specified distance by returning a new dancer state.

SquareDanceReasoning.rotate — Method

rotate(ds::DancerState, rotation, time_delta)::DancerState

Rotates the dancer idntified by the DancerState in place by rotation.

SquareDanceReasoning.schedule — Method

schedule(sched::CallSchedule, call::SquareDanceCall, at)

Enters call into sched so that it will be performed at the specified time at.

SquareDanceReasoning.showcase — Method

showcase(filename::String, title::String, initial_kb, choreography::Vector{SquareDanceCall}; inhibit_call_engine_logging = true; dbgctx = nothing)

Writes a markdown file named filename in the choreography Showcase documentation directory.

title is the title string for the document.

initial_kb is the knowledgebase containing the formations that choreography starts from.

choreography is a vector of the calls to be performed.

If inhibit_call_engine_logging is true then teh usual call engine logging will be suppressed.

If provided, dbgctx should be a CallEngineDebugContext.

SquareDanceReasoning.square_up — Method

square_up(dancers; initial_time = 0)

returns a list of DancerStates for the initial squared set.

SquareDanceReasoning.step_to_a_wave — Method

step_to_a_wave(f::FaceToFace, time_delta, h::Handedness)::MiniWave

The face to face dancers move up to make a right or left handed miniwave.

Breating should be done separately.

SquareDanceReasoning.those_with_role — Method

those_with_role(::SquareDanceFormation, ::Role)

For the specified formation, returns DancerStates or subformations whose dancers fill the specified role.

SquareDanceReasoning.un_step_to_a_wave — Method

un_step_to_a_wave(f::MiniWave, time_delta)::FaceToFace

The face to face dancers move up to make a right or left handed miniwave.

Breating should be done separately.

SquareDanceReasoning.update_from — Function

update_from(formation::SquareDanceFormation, newest::Vector{DancerState})

Return a new formation of the same type as formation but with its component DancerStates taken from newest, matching by Dancer.

SquareDanceReasoning.@CallEngineDebugContext — Macro

@CallEngineDebugContext(relpath, token)

Creates a CallEngineDebugContext with an appropriate source locator.